+++ src/main/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: de.fraunhofer.aisec.cpg.ptn4j.Application
+Main-Class: de.fraunhofer.aisec.cpg_vis_neo4j.Application
